% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_asinh.R
\name{asinhTransform}
\alias{asinhTransform}
\title{Create a FlowCore Compatible Transform Object}
\usage{
asinhTransform(transformId = "myasinh", cofactor = 150, jitter = TRUE)
}
\arguments{
\item{transformId}{FlowCore compatible transformId}

\item{jitter}{Logical, should small values be jittered for visual appeal?}

\item{a}{The "a" parameter.  Do not override unless you know what you're doing.}

\item{full_scale}{The value of full-scale events (default = 2^18 - 1)}
}
\value{
an object of type \code{transform}, to be used with flowCore functions.
}
\description{
Create a flowCore compatible transform object.

This function creates a flowCore compatible transform object.
}
