% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_asinh.R
\name{asx}
\alias{asx}
\title{Shorthand Function for Arcsinh Transformation}
\usage{
asx(x, cofactor = 5)
}
\arguments{
\item{x}{A numeric scalar or vector to be transformed}

\item{cofactor}{Adjustable parameter. Default = 5.}
}
\value{
Transformed values.
}
\description{
A shorthand function to perform the arcsinh transform on numeric values
}
\seealso{
\code{\link{ibx}}
}
