% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_autocomp.R
\name{autocomp}
\alias{autocomp}
\title{Apply a Compensation Matrix}
\usage{
autocomp(f)
}
\arguments{
\item{f}{The input flowFrame or flowSet}
}
\value{
A compensated flowFrame or flowSet
}
\description{
Apply the compensation matrix in the FCS header.
}
