% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_biexp.R
\name{biexp}
\alias{biexp}
\title{Calculate a Biexponential Transformation}
\usage{
biexp(ff, a = 0.002, params)
}
\arguments{
\item{ff}{a flowFrame}

\item{a}{The "a" parameter.  Do not override unless you know what you're doing.}

\item{params}{The parameters of ff to transform}
}
\value{
A transformed flowFrame
}
\description{
Do the biexponential mathematical transform

This function performs the biexponential transform directly on ff.
See \code{\link{biexpTransform}} for flowCore compatibility.
}
