% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_biexp.R
\name{biexpTransform}
\alias{biexpTransform}
\title{Create a FlowCore Compatible Transform Object}
\usage{
biexpTransform(
  transformId = "mybiexp",
  a = 0.002,
  full_scale = 262143,
  jitter = TRUE
)
}
\arguments{
\item{transformId}{FlowCore compatible transformId}

\item{a}{The "a" parameter.  Do not override unless you know what you're doing.}

\item{full_scale}{The value of full-scale events (default = 2^18 - 1)}

\item{jitter}{Logical, should small values be jittered for visual appeal?}
}
\value{
an object of type \code{transform}, to be used with flowCore functions.
}
\description{
Create a flowCore compatible transform object.

This function creates a flowCore compatible transform object.
}
