% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_boxcox.R
\name{boxcoxTransform}
\alias{boxcoxTransform}
\title{Create a FlowCore Compatible Transform Object}
\usage{
boxcoxTransform(
  transformId = "myboxcox",
  lambda = 0.1,
  scale = NA,
  jitter = FALSE
)
}
\arguments{
\item{transformId}{FlowCore compatible transformId}

\item{lambda}{boxcox lambda parameter}

\item{scale}{boxcox scale parameter}

\item{jitter}{Logical, should small values be jittered for visual appeal?}
}
\value{
an object of type \code{transform}, to be used with flowCore functions.
}
\description{
Do the boxcox mathematical transform

This function creates a flowCore compatible transform object.
}
