% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_biexp.R
\name{bx}
\alias{bx}
\title{Shorthand Function for Biexponential Transformation}
\usage{
bx(x)
}
\arguments{
\item{x}{A numeric scalar or vector to be transformed}
}
\value{
Transformed values.
}
\description{
A shorthand function to perform the biexponential transform on numeric values
}
\seealso{
\code{\link{ibx}}
}
