% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{centroid}
\alias{centroid}
\title{Centroid of a Blob Contour}
\usage{
centroid(cont)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()
the smoother the result.}
}
\value{
The 2-dimensional blob centroid.
}
\description{
Find the center of a blob, defined as the mean of the vertices
of the blob.
}
