% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_compute_spill.R
\name{compute.spill}
\alias{compute.spill}
\title{Calculate a Spillover Matrix from Diva Controls}
\usage{
compute.spill(ff, fsc.thresh = 30000, path = NULL, show = FALSE)
}
\arguments{
\item{ff}{A flowFrame, used to specify fluorescence parameter names.  \strong{Please note,
the parameters in ff MUST be original - that is, there must not have been a
"nice.names" or "swap.names" maneuver prior to calling this function!}}

\item{fsc.thresh}{A threshold to be applied on the FSC-A parameter, to
eliminate debris}

\item{path}{The directory containing the comp control tubes.  If not given,
path is assumed to be the current working directory.}

\item{show}{Logical, should we show our work?}
}
\value{
The calculated spillover matrix.
}
\description{
Calculate a spillover matrix, given compensation control FCS
files in DIVA convention.
}
\details{
This function computes a spillover matrix using the data in a collection
of compensation control FCS files.  These files MUST reside in the same directory.
We assume that the acquisition program was Diva.  Other vendors/conventions are
not currently supported.

There is an (optional) pre-gating step on FSC-A, designed to eliminate debris.  The
default value for this parameter assumes no adjustment (normalization) of the
original values.

The method involves doing linear regression on the spill data, to derive the
slope of the data of y versus x, where x is the stained parameter, and y is all
other parameters.  Regression occurs iteratively, eliminating outliers in an
attempt to find a robust slope.

Please note that setting \code{show = TRUE} will be very slow, and with large numbers
of parameters may overwhelm your graphics capability.
}
