% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_histogram_analysis.R
\name{cumm.kde}
\alias{cumm.kde}
\title{Calculate the Cummulative Probability Distribution of a KDE}
\usage{
cumm.kde(kde)
}
\arguments{
\item{kde}{The input kde}
}
\value{
A list with 2 elements:
\describe{
\item{x}{X coordinates of the CDF}
\item{y}{Y coordinates of the CDF, normalized to 1.0.}
}
}
\description{
The cummulative distribution function (think of the kde as a
differential probability distribuion).
}
