% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_deGate.R
\name{deGate}
\alias{deGate}
\title{A Fancy 1D Threshold Calculator}
\usage{
deGate(
  f,
  channel,
  n.sd = 1.5,
  high = FALSE,
  percentile,
  kernel = "gaussian",
  graphs = FALSE,
  all.cut = FALSE
)
}
\arguments{
\item{f}{A flowFrame}

\item{channel}{an integer to specifiy the channel for 1D density estimation analysis}

\item{n.sd}{an integer that is multiplied to the standard deviation to determine the place of threshold}

\item{high}{if TRUE, returns the 'percentile' threshold. It returns the 99th percentile by default}

\item{percentile}{a value in the interval 0,1 that is used as the percentile is 'high' is TRUE}

\item{kernel}{refer to the '?density' in r base}

\item{graphs}{if TRUE, it plots the density as well as the threshold on the same plot}

\item{all.cut}{if TRUE, it returns all the cutoff points whose length can roughly estiamte the number of cell subsets in that dimension}
}
\value{
cutoffs, i.e. thresholds on the 2D data
}
\description{
Estimates the number of cell subsets and identifies the best threshold to gate the positive/negative subsets of cell populations in 1D.
Note:  This function was written by M. Jafar Taghiyar and included with his kind permission.
}
