% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cleanup.R
\name{deneg}
\alias{deneg}
\title{Get Rid of Events Whose Parameter Value is Below Zero}
\usage{
deneg(ff, parameters = NULL, min.value = 0)
}
\arguments{
\item{ff}{The input flowFrame}

\item{parameters}{The parameters of ff you wish to derail}

\item{min.value}{The value below which events will be removed from the flowFrame}
}
\value{
A de-neg'd flowFrame.
}
\description{
Eliminate values less than min.value from a flowFrame.
}
