% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cleanup.R
\name{derail}
\alias{derail}
\title{Get Rid of Events on the Rail}
\usage{
derail(ff, parameters = NULL, fac = 0.99)
}
\arguments{
\item{ff}{The input flowFrame}

\item{parameters}{The parameters of ff you wish to derail}

\item{fac}{A factor < 1.0 to use as a 'buffer zone'}
}
\value{
A de-railed flowFrame.
}
\description{
Derail (that is, remove events on the positive axis) of selected
parameters
}
\details{
derail consults the \strong{maxRange} values in the phenodata slot of the
flowFrame ff.  If something has gone wrong, it's likely due to the transformation
or normalization of data in the flowFrame.
}
