% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_histogram_analysis.R
\name{deriv1.kde}
\alias{deriv1.kde}
\title{Estimate the First Derivative of a KDE}
\usage{
deriv1.kde(kde, normalize = TRUE)
}
\arguments{
\item{kde}{The input kde}

\item{normalize}{Logical, should the resulting kde be normalized to 1.0?}
}
\value{
Another kde.
}
\description{
Estimate the first derivative of a kde using the symmetric difference quotient.
}
