% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_importutils.R
\name{detect_fl_parameters}
\alias{detect_fl_parameters}
\title{Detect Fluorescence Parameters in a flowFrame}
\usage{
detect_fl_parameters(ff, non.fl = c("fsc", "ssc", "time", "clean", "index"))
}
\arguments{
\item{ff}{The flowFrame to be processed}

\item{non.fl}{Names of all non-fluorescence parameters in the flowFrame}
}
\value{
The indices corresponding to the scattering parameters.
}
\description{
Return the indices of the fluorescence parameters.
}
\details{
This function assumes that FL parameters are anything that don't contain
(non-case-sensitive) "fsc", "ssc", "time", "clean" or "index", or any other
parameter names supplied.
in the argument, non.fl.
}
