% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_importutils.R
\name{detect_scatter_parameters}
\alias{detect_scatter_parameters}
\title{Detect Scattering Parameters in a flowFrame}
\usage{
detect_scatter_parameters(ff)
}
\arguments{
\item{ff}{The flowFrame to be processed}
}
\value{
The indices corresponding to the scattering parameters.
}
\description{
Return the indices of the scattering parameters.
}
\details{
This function will work with files whose scattering parameter names
include "FSC" or "SSC".  Not all instruments name parameters thusly.
}
