% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_importutils.R
\name{doTransform}
\alias{doTransform}
\title{A Convenience Wrapper for Transformation}
\usage{
doTransform(
  ff,
  cols = c(1:5, 7:13),
  method = c("biexp", "asinh", "log", "linear"),
  fac = 5.4/262143
)
}
\arguments{
\item{ff}{The flowFrame to be transformed}

\item{cols}{The columns (parameters) of the flowFrame to be transformed}

\item{method}{The transformation function. One of \code{biexp}, \code{log} or \code{linear}.}

\item{fac}{The scaling factor used only for \code{method = 'linear'}.}
}
\value{
The transformed flowFrame.
}
\description{
Convenience wrapper for \code{\link[flowCore]{transformList}} and \code{\link[flowCore]{transform}}.
}
\details{
Please refer to the code for the math details.  The biexp method is specific to this
package.  It's similar to other biexponential transforms (e.g. flowCore, FlowJo) but differs
in detail.
}
