% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_histogram_analysis.R
\name{find.local.maxima}
\alias{find.local.maxima}
\title{Find Local Maxima in a KDE}
\usage{
find.local.maxima(kde, thresh = 0.05, show = FALSE, ...)
}
\arguments{
\item{kde}{The input Kernel Density Estimate}

\item{thresh}{Threshold above which peaks are considered}

\item{show}{Logical, should we display our work?
#' @param ... Additional arguments passed to par().  Relevant only if show = TRUE.}
}
\value{
A list with three elements:
\describe{
\item{x}{X coordinates of the detected peaks}
\item{y}{Y coordinates of the detected peaks}
\item{pick}{The indices in the original kde corresponding to the peaks}
}
}
\description{
Find the peaks in a Kernel Density Estimate.
}
