% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_histogram_analysis.R
\name{find.width}
\alias{find.width}
\title{Calculate the Width of a KDE}
\usage{
find.width(kde, frac = 0.5)
}
\arguments{
\item{kde}{The input kde}

\item{frac}{The height (\link{0, 1}) at which to compute the width.}
}
\value{
A list with 4 elements:
\describe{
\item{center}{X coordinate of the peak center}
\item{left}{X coordinate of the left intercept}
\item{right}{X coordinate of the right intercept}
\item{frac}{The value of the input parameter 'frac'}
}
}
\description{
find the full width half-maximum (or other width at percent maximum) of a kde.
}
