% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gaussian_utils.R
\name{fit_gaussian_base}
\alias{fit_gaussian_base}
\title{Generate a Gaussian From a KDE}
\usage{
fit_gaussian_base(kde, height, show = FALSE)
}
\arguments{
\item{kde}{The input kde}

\item{height}{The height at which the generated gde}

\item{show}{Logical.  Should we show our work?}
}
\value{
A gde.
}
\description{
Generate a gaussian curve, in the same form as our familiar
kde's (a 'gde'), from an input kde.  The resulting gaussian  intercepts the
input kde at the given height.
}
