% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gaussian_utils.R
\name{gaussian_crossover}
\alias{gaussian_crossover}
\title{Find the Crossover Point of Two Gaussians}
\usage{
gaussian_crossover(l1, s1, m1, l2, s2, m2)
}
\arguments{
\item{l1}{Lambda of the first gaussian}

\item{s1}{Sigma of the first gaussian}

\item{m1}{Mu of the first gaussian}

\item{l2}{Lambda of the second gaussian}

\item{s2}{Sigma of the second gaussian}

\item{m2}{Mu of the second gaussian}
}
\value{
The x-coordinate of the intersection.
}
\description{
Find the crossover of 2 gaussians, that is, where they intersect.
}
