% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gaussian_utils.R
\name{gaussian_dist}
\alias{gaussian_dist}
\title{Generate a Gaussian GDE Distribution}
\usage{
gaussian_dist(lambda, sigma, mu, x = seq(-1, 5, length.out = 200))
}
\arguments{
\item{lambda}{Lambda}

\item{sigma}{Sigma}

\item{mu}{Mu}

\item{x}{The x values at which the gaussian densities will be generated}
}
\value{
A list with 2 elements:
\describe{
\item{x}{The x values of the gde}
\item{y}{The y values of the gde}
}
}
\description{
Generate a Gaussian Density Function (gde) given mu, sigma and lambda.
}
