% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{get.hull}
\alias{get.hull}
\title{Get the Convex Hull of a Blob}
\usage{
get.hull(blob)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()}
}
\value{
Another polygon
}
\description{
find the convex hull of a polygon
}
