% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_asinh.R
\name{iasx}
\alias{iasx}
\title{Shorthand Function for Inverse Arcsinh Transformation}
\usage{
iasx(x, cofactor = 150)
}
\arguments{
\item{x}{A numeric scalar or vector to be transformed}
}
\value{
Transformed values.
}
\description{
A shorthand function to perform the inverse arcsinh transform on numeric values
}
\seealso{
\code{\link{bx}}
}
