% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_biexp.R
\name{ibx}
\alias{ibx}
\title{Shorthand Function for Inverse Biexponential Transformation}
\usage{
ibx(x)
}
\arguments{
\item{x}{A numeric scalar or vector to be transformed}
}
\value{
Transformed values.
}
\description{
A shorthand function to perform the inverse biexponential transform on numeric values
}
\seealso{
\code{\link{bx}}
}
