% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{inflate.contour}
\alias{inflate.contour}
\title{Inflate a Blob Contour}
\usage{
inflate.contour(blob, dist)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()}

\item{dist}{The distance by which to expand the blob}
}
\value{
Another polygon
}
\description{
Expand the input blob radially by a constant distance.
}
\details{
The input blob should be convex.  You can use get.hull() first if you're not sure.
}
