% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_biexp.R
\name{inv.biexp}
\alias{inv.biexp}
\title{Calculate an Inverse Biexponential Transformation}
\usage{
inv.biexp(ff, a = 0.002, params)
}
\arguments{
\item{ff}{a flowFrame}

\item{a}{The "a" parameter.  Do not override unless you know what you're doing.}

\item{params}{The parameters of ff to transform}
}
\value{
A transformed flowFrame
}
\description{
Do the inverse biexponential mathematical transform

This function performs the inverse biexponential transform directly on ff.
}
