% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_histogram_analysis.R
\name{normalize.kde}
\alias{normalize.kde}
\title{Normalize a KDE}
\usage{
normalize.kde(kde)
}
\arguments{
\item{kde}{The input kde}
}
\value{
A normalized version of the kde.
}
\description{
Normalize a kde to 1.0
}
