% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{polygon.area}
\alias{polygon.area}
\title{Calculate a Blob's Area}
\usage{
polygon.area(blob)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()}
}
\value{
The area of the polygon.
}
\description{
Calculate the area of a blob (polygon).
}
\details{
See http://stackoverflow.com/questions/16285134/calculating-polygon-area.
}
