% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{polygon.convexity}
\alias{polygon.convexity}
\title{Calculate a Blob's Convexity}
\usage{
polygon.convexity(blob)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()}
}
\value{
The convexity of the polygon.
}
\description{
Calculate the convexity of a blob(polygon).
}
\details{
Convexity is defined as the ratio of the area of a polygon to the area
of its convex hull.  A convex polygon thus has a convexity of 1.0.
}
