% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pplot.R
\name{pplot}
\alias{pplot}
\title{Draw a FlowJo-style (sort of) Plot}
\usage{
pplot(
  ff,
  parameters,
  blueBackground = FALSE,
  showZero = FALSE,
  nbin = 501,
  bandwidth = 0.001,
  cr = blob_color(blueBackground = FALSE),
  col = "black",
  nrpoints = 0,
  max_channel = 262143,
  tx = c("biexp", "log", "linear"),
  ty = c("biexp", "log", "linear"),
  plotaxt = TRUE,
  ticksize = 1,
  ...
)
}
\arguments{
\item{ff}{The flowFrame to be plotted}

\item{parameters}{A vector of two parameters in ff.  May be either character or numeric}

\item{blueBackground}{Logical, should the background be blue.  If false, you'll get a white
background (like FlowJo)}

\item{showZero}{Logical, draw dot-dash lines indicating the location of zero on each axis}

\item{nbin}{Binning resolution (you shouldn't have to mess with the default)}

\item{bandwidth}{Binning bandwidth (you shouldn't have to mess with the default)}

\item{cr}{Color ramp.  The only one supplied currently is the default}

\item{col}{Not sure about this one, but the default is ok}

\item{nrpoints}{Not sure about this one, but the default is ok}

\item{max_channel}{Max value the cytometer can produce.  See \code{\link{ax}}}

\item{tx}{tick style.  One of \code{'biexp', 'log', 'linear'}}

\item{ty}{tick style.  One of \code{'biexp', 'log', 'linear'}}

\item{plotaxt}{Logical, should we draw the axes?}

\item{ticksize}{Weight of the tickmarks in relative coordinates.}

\item{...}{Additional graphical parameters passed to \code{flowCore \link[flowCore]{plot}}}
}
\description{
Draw a picture, with colors reminiscent of FlowJo.
}
\examples{

# a plot with linear coordinate scaling
pplot(ff, parameters = c("FSC-A", "SSC-A"), tx = 'linear', ty = 'linear')

# another plot, with the default biexponential scaling
pplot(ff, c("CD4", "CD8"))
}
