% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_importutils.R
\name{replace_phedata_ranges}
\alias{replace_phedata_ranges}
\title{Replace range data in the pData(parameters(ff)) slot}
\usage{
replace_phedata_ranges(ff, range = NULL, minRange = NULL, maxRange = NULL)
}
\arguments{
\item{ff}{The flowFrame to be processed}

\item{range}{The range values to replace those in ff}

\item{minRange}{The minRange values to replace those in ff}

\item{maxRange}{The maxRange values to replace those in ff}
}
\value{
A flowFrame with replace range values.
}
\description{
We found that round-tripping a flowFrame by first \code{ \link[flowCore]{write.FCS}}
followed by \code{ \link[flowCore]{read.FCS}} caused weird effects in the phenodata
of the resulting flowFrame.  This function replaces the screwed up range numbers
with ones stored from before the read-back.
}
\details{
This function should be used with care.  The range values are affected
by the per-parameter transformations that may or may not have been previously
applied.  There is currently no good mechanism to track them.

The default values of NULL for range, minRange and maxRange result in the generation
of values appropriate for data that have been (a) linearly scaled to 0-5.4
on scattering parameters, and (b) biexponentially transformed (using \code{ \link{biexpTransform}})
on the fluorescence parameters.  Otherwise, care should be taken to supply
range, minRange and maxRange vectors having the same length as ncol(ff).
}
