% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gaussian_utils.R
\name{retrieve_sigma_mu}
\alias{retrieve_sigma_mu}
\title{Get the Gaussian Parameters of a GDE}
\usage{
retrieve_sigma_mu(gde)
}
\arguments{
\item{gde}{The input Gaussian Density Estimate}
}
\value{
A list with 2 elements:
describe{
\item{sigma}{The sigma value for the input gde}
\item{mu}{ The mu value for the input gde}
}
}
\description{
Find sigma, mu for a given Gaussian Density Estimate
(see for example \code{\link{fit_gaussian_base}})
}
