% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_blob_utils.R
\name{smooth.contour}
\alias{smooth.contour}
\title{Smooth a Blob Contour}
\usage{
smooth.contour(blob, npts = 5)
}
\arguments{
\item{blob}{A polygon, for example the result of blob.boundary()}

\item{npts}{The number of points used in smoothing.  The larger the number
the smoother the result.}
}
\value{
Another polygon
}
\description{
Create a smoother representation of the input blob.
}
