% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_textutils.R
\name{text_to_path}
\alias{text_to_path}
\title{Make a Properly Formatted File Path}
\usage{
text_to_path(...)
}
\arguments{
\item{...}{One or more character scalars to be concatenated with \code{/} between elements.}
}
\value{
A character variable containing the concatenated strings, properly delimited with
slashes.
}
\description{
Concatenate character elements into a unix-like path specification.
}
\examples{

a = "dir1"
b = "dir2"
c = "dir3"
text_to_path(a, b, c)
## "dir1/dir2/dir3/"

}
