% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_axes.R
\name{tick.labels}
\alias{tick.labels}
\title{Label (Quasi-)Logarithmic Tickmarks}
\usage{
tick.labels(first_decade, last_decade)
}
\arguments{
\item{first_decade}{First decade (e.g. 1 would correspond to 10^1)}

\item{last_decade}{Last decade (e.g. 5 would correspond to 10^5)}
}
\value{
A list of parsed labels suitable for the "label" parameter of \link{ax}
or \link[graphics]{axis}.
}
\description{
Create powers-of-ten tickmark labels
}
