% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_textutils.R
\name{tight}
\alias{tight}
\title{Paste Tightly}
\usage{
tight(...)
}
\arguments{
\item{...}{One or more character scalars to be pasted tightly.}
}
\value{
A character variable containing the concatenated strings.
}
\description{
Paste tightly, without having to specify the \code{sep} argument.
}
\examples{
\code{
a = "The_first_part"
b = "---"
c = "and_the_second_part"
tight(a, b, c)
## "The_first_part---and_the_second_part"

}
}
