% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_asinh.R
\name{w.inv.arcsinh}
\alias{w.inv.arcsinh}
\title{Calculate an Inverse Arcsinh Transformation}
\usage{
w.inv.arcsinh(ff, cofactor = 150, params)
}
\arguments{
\item{ff}{a flowFrame}

\item{params}{The parameters of ff to transform}

\item{a}{The "a" parameter.  Do not override unless you know what you're doing.}
}
\value{
A transformed flowFrame
}
\description{
Do the inverse arcsinh mathematical transform

This function performs the inverse arcsinh transform directly on ff.
}
