% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_importutils.R
\name{doTransform}
\alias{doTransform}
\title{A Convenience Wrapper for Transformation}
\usage{
doTransform(
  ff,
  cols = c(1:5, 7:13),
  method = c("biexp", "asinh", "log", "linear"),
  a = 0.002,
  cofactor = 5,
  fac = 5.4/262143
)
}
\arguments{
\item{ff}{The flowFrame to be transformed}

\item{cols}{The columns (parameters) of the flowFrame to be transformed}

\item{method}{The transformation function. One of \code{biexp}, \code{log} or \code{linear}.}

\item{a}{The adjustable parameter for Wade's biexpTransform (Default = 0.002)}

\item{cofactor}{The adjustable parameter for Wade's asinhTransform (Default = 5)}

\item{fac}{The scaling factor used only for \code{method = 'linear'}.}
}
\value{
The transformed flowFrame.
}
\description{
Convenience wrapper for \code{\link[flowCore]{transformList}} and \code{\link[flowCore]{transform}}.
}
\details{
Please refer to the code for the math details.  The biexp and asinh methods are specific to this
package.  They're similar to other corresponding transforms (e.g. flowCore, FlowJo) but differ
in detail.  The values of the two adjustable transformation parameters
\code{a} and \code{cofactor} are only applicable to their corresponding transforms.
}
